Fixing code structure removing unused flags adding more error checks
