--- Coreutils/echo.go
@@ -1,14 +1,0 @@
-package coreutils
-
-import (
-	"fmt"
-	"os"
-)
-
-func Echo(n bool) {
-	argsWithProg := os.Args
-	if n {
-		argsWithProg = os.Args[1:]
-	}
-	fmt.Println(argsWithProg)
-}
