+++ Coreutils/false/main.go
@@ -1,0 +1,11 @@
+package main
+
+import "fmt"
+
+func False() int {
+	return 0
+}
+
+func main() {
+	fmt.Println(False())
+}
