+++ Coreutils/head/input.txt
@@ -1,0 +1,12 @@
+Here is start
+Here is EXample
+Here is EXample
+Here is EXample
+Here is EXample
+Here is EXample
+Here is EXample
+Here is EXample
+Here is EXample
+Here is EXample
+Here is EXample
+Here is tail
