+++ go.mod
@@ -1,3 +1,3 @@
-module Coreutils
+module github.com/codescalersinternships/Coreutils-Nabil
 
 go 1.21.6
