+++ readme.md
@@ -8,21 +8,21 @@ Goal is to reimplement several core Unix utilities.
 
 The `head` command in Go prints the first N lines of a file. Implemented with:
 ```bash
-go run main.go -operation=head -directory=input.txt -isFile=true -n=10
+go run ./head -n=2 input.txt
 ```
 
 ### 2. tail
 
 The `tail` command in Go prints the last N lines of a file. Implemented with:
 ```bash
-go run main.go -operation=tail -directory=input.txt -isFile=true -n=10
+go run ./tail -n=2 input.txt
 ```
 
 ### 3. wc
 
 The `wc` command in Go counts lines, words, and bytes in a file and prints the counts.
 ```bash
-go run main.go -operation=wc -directory=input.txt -isFile=true -l=true -w=true -c=true
+go run ./wc -l=true -w=true -c=true input.txt
 ```
 
 ### 4. cat
@@ -30,40 +30,40 @@ go run main.go -operation=wc -directory=input.txt -isFile=true -l=true -w=true -
 The `cat` command in Go concatenates files and prints them to standard output. Implemented with:
 - `cat -n`: Prints each line with line numbers.
 ```bash
-go run main.go -operation=cat -directory=input.txt
+go run ./cat -n input.txt
 ```
 
 ### 5. echo
 
 The `echo` command in Go prints its arguments to standard output.
 ```bash
-go run main.go -operation=echo -directory=input.txt -0=true
+go run ./echo -n=true hello
 ```
 
 ### 6. tree
 
 The `tree` command in Go recursively lists contents of directories in a tree-like format.
 ```bash
-go run main.go -operation=tree -directory=input.txt
+go run ./tree ./
 ```
 
 ### 7. env
 
 The `env` command in Go lists the current environment variables and their values.
 ```bash
-go run main.go -operation=env -directory=input.txt
+go run ./env
 ```
 
 ### 8. true
 
 The `true` command in Go does nothing except return a successful exit status.
 ```bash
-go run main.go -operation=true
+go run ./true
 ```
 
 ### 9. false
 
 The `false` command in Go does nothing except return a non-successful exit status.
 ```bash
-go run main.go -operation=false
+go run ./false
 ```
